LayoutTests/platform/ios/ios/fast/forms/autocapitalize-null.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../../../resources/js-test-pre.js"></script>
<script>
description("Test null handling on the autocapitalize attribute");

var form = document.createElement("form");
shouldBeNull("form.getAttribute('autocapitalize')");
evalAndLog("form.autocapitalize = null");
shouldBeEqualToString("form.getAttribute('autocapitalize')", "");
shouldBeEqualToString("form.autocapitalize", "");

var input = document.createElement("input");
shouldBeNull("input.getAttribute('autocapitalize')");
evalAndLog("input.autocapitalize = null");
shouldBeEqualToString("input.getAttribute('autocapitalize')", "");
shouldBeEqualToString("input.autocapitalize", "");

var textarea = document.createElement("textarea");
shouldBeNull("textarea.getAttribute('autocapitalize')");
evalAndLog("textarea.autocapitalize = null");
shouldBeEqualToString("textarea.getAttribute('autocapitalize')", "");
shouldBeEqualToString("textarea.autocapitalize", "");
</script>
<script src="../../../../../resources/js-test-post.js"></script>
</body>
</html>
